+++ metrics/cli/lib/cli/doctor/doctor.dart
@@ -1,7 +1,7 @@
 // Use of this source code is governed by the Apache License, Version 2.0
 // that can be found in the LICENSE file.
 import 'package:cli/common/model/services.dart';
-import 'package:cli/services/common/service/info_service.dart';
+import 'package:cli/services/common/info_service.dart';
 import 'package:cli/services/firebase/firebase_service.dart';
 import 'package:cli/services/flutter/flutter_service.dart';
 import 'package:cli/services/gcloud/gcloud_service.dart';
